==> Code/6_Wait_Time_Analysis.Rmd <==
---
title: "Wait Time Analysis"
author: "Clare Clingain"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
require(lubridate)
require(dplyr)
library(tidyverse)
load("../Data/Clean/analysis_data.RData")
```

# Research Questions

Here are the research questions I hope to answer:

1) What does the distribution of lateness look like for connections between the 1 train and the Staten Island Ferry?

2) How does this distribution vary across time (e.g., rush hour vs tourist peak)?

3) How does this distribution vary for people who can run versus people who can't run from the 1 train to the ferry?

# Initial Visualizations

## Density plots

```{r density}
# overlay plot
plot(density(analysis_data$wait_times_1min), main = "Wait times with 1, 2, and 3 minutes exit time", col = 2)
lines(density(analysis_data$wait_times_2min), col = 4)
lines(density(analysis_data$wait_times_3min), col = 5)
legend("topright",legend = c("1 min","2 min","3 min"), lty = 1, lwd = 2, col = c(2,4,5))
```

The three distributions look quite similar. In terms of the simulations I ran, these distributions look like the beta distribution results. 

## Time (hour) x Wait times

```{r time by wait times}
plot(as.factor(analysis_data$hour), analysis_data$wait_times_1min, main = "Wait times  (1 min exit time)",xlab = "Hour of day",ylab = "Wait time")
plot(as.factor(analysis_data$hour), analysis_data$wait_times_2min, main = "Wait times  (2 min exit time)",xlab = "Hour of day",ylab = "Wait time")
plot(as.factor(analysis_data$hour), analysis_data$wait_times_3min, main = "Wait times (3 min exit time)",xlab = "Hour of day",ylab = "Wait time")
```

It makes sense that the waits pile up late at night and during the tourist peak hours (10-3). I'm curious seeing that the very lengthy wait times caused by a missing ferry are so dispersed. I initially thought that it would be worse in the wee hours of the morning, but it looks to be a problem throughout the day. This isn't concerning in terms of something being wrong with the data because there are many reasons why a ferry or a train never shows. Both can break down, both can have sick passengers, and both can barely miss each other. 

```{r med iqr}
# medians
medians_IQR <- analysis_data %>% 
  group_by(as.factor(hour)) %>% 
  summarise(median_1min = median(wait_times_1min),
            median_2min = median(wait_times_2min),
            median_3min = median(wait_times_3min),
            IQR_1min = IQR(wait_times_1min),
            IQR_2min = IQR(wait_times_2min),
            IQR_3min = IQR(wait_times_3min))
colnames(medians_IQR)[1] <- "hour"
# plot the medians and IQR
plot(x = as.numeric(medians_IQR$hour), y = medians_IQR$median_1min, pch = 16, main = "Median wait times by hour", ylab = "Wait time (in minutes)", xlab = "Hour")
lines(x = as.numeric(medians_IQR$hour), y = medians_IQR$median_1min, lwd = 2)
lines(x = as.numeric(medians_IQR$hour), y = medians_IQR$median_2min, lwd = 2, col = 2, type = 'o', pch = 16)
lines(x = as.numeric(medians_IQR$hour), y = medians_IQR$median_3min, lwd = 2, col = 4, type = 'o', pch = 16)
legend("bottomright",legend = c("1 min","2 min","3 min"), lty = 1, lwd = 2, col = c(1,2,4))
```

For the rush-hour and tourist peak hours, it looks like time-to-exit doesn't make a difference. A longer time-to-exit does seem to matter for the late night/early morning hours. 

## Time (weekday) x Wait times

Now I'll check the wait times by weekday. 

```{r weekday by wait}
plot(analysis_data$weekday, analysis_data$wait_times_1min, main = "Wait times (1 min exit time)",xlab = "Day of the week",ylab = "Wait time")
plot(analysis_data$weekday, analysis_data$wait_times_2min, main = "Wait times (2 min exit time)",xlab = "Day of the week",ylab = "Wait time")
plot(analysis_data$weekday, analysis_data$wait_times_3min, main = "Wait times (3 min exit time)",xlab = "Day of the week",ylab = "Wait time")
```

There doesn't seem to be much difference across days of the week. 

## Time (Month) x Wait times

```{r month wait time}
plot(as.factor(analysis_data$month), analysis_data$wait_times_1min, main = "Wait times (1 min exit time)",xlab = "Month",ylab = "Wait time")
plot(as.factor(analysis_data$month), analysis_data$wait_times_2min, main = "Wait times (2 min exit time)",xlab = "Month",ylab = "Wait time")
plot(as.factor(analysis_data$month), analysis_data$wait_times_3min, main = "Wait times (3 min exit time)",xlab = "Month",ylab = "Wait time")
```

## Time (Rush vs nonrush) x Wait times

Next, I'll have a look at the three distributions by rush hour. 

```{r rush wait}
plot(as.factor(analysis_data$rushour), analysis_data$wait_times_1min, main = "Wait times (1 min exit time)",xlab = "Rush hour vs Non-rush hour",ylab = "Wait time")
plot(as.factor(analysis_data$rushour), analysis_data$wait_times_2min, main = "Wait times (2 min exit time)",xlab = "Rush hour vs Non-rush hour",ylab = "Wait time")
plot(as.factor(analysis_data$rushour), analysis_data$wait_times_3min, main = "Wait times (3 min exit time)",xlab = "Rush hour vs Non-rush hourk",ylab = "Wait time")
```

As expected, wait times are longer outside of rush hour. However, non rush hour includes late night hours as well as tourist peak hours. I'll break it down further by tourist peak hours. 

# Test the difference between the wait distributions

I will use the two-sample Kolmogorov-Smirnov test to determine whether each distribution is different from the other two. I will adjust my p-values to reflect the multiple tests. 

```{r kolmogorov smirnov}
one_two_res <- ks.test(analysis_data$wait_times_1min, analysis_data$wait_times_2min)
one_two_res
one_three_res <- ks.test(analysis_data$wait_times_1min, analysis_data$wait_times_3min)
one_three_res
two_three_res <- ks.test(analysis_data$wait_times_2min, analysis_data$wait_times_3min)
two_three_res

# adjust p-values
paste("The adjusted p-values for each respective test are",p.adjust(c(one_two_res$p.value, one_three_res$p.value, two_three_res$p.value), method = "hochberg"))
```

We see a significant difference between all three distributions. Although this is great news, I strongly wonder if the data violate the iid assumption of the KS test. Even though some wait times may be independent of one another (i.e., morning rush from evening rush, January from June), I doubt that this is the case across all the data. Two consectutive trains are unlikely to be independent. Nevertheless, this test gives us an additional piece of information that we will use with caution. 

# Test difference between distribution means

As discussed above, the iid assumption is trick with this data. However, I'd like to get some 95% confidence intervals for the difference between the means.

```{r t tests}
t1 <- t.test(analysis_data$wait_times_1min, analysis_data$wait_times_2min)
t2 <- t.test(analysis_data$wait_times_1min, analysis_data$wait_times_3min)
t3 <- t.test(analysis_data$wait_times_2min, analysis_data$wait_times_3min)
paste("The 95% confidence interval for the difference in means between 1min vs 2min is (", round(t1$conf.int[1],3),",", round(t1$conf.int[2],3),")")
paste("The 95% confidence interval for the difference in means between 1min vs 3min is (", round(t2$conf.int[1],3),",", round(t2$conf.int[2],3),")")
paste("The 95% confidence interval for the difference in means between 2min vs 3min is (", round(t3$conf.int[1],3),",", round(t3$conf.int[2],3),")")
```

# Probality of Wait Times

Next I will calculate the probability of waiting 27-29 minutes (the myth), of waiting 30+ minutes (nightmare commute), and of waiting less than 5 minutes for each of the distributions.

## The 1 minute myth

I'm allowing for a range of 1 to 3 minutes to qualify as the myth given that I'm looking at a range of exit times. First I'll calculate the marginal probabilities, and then move onto conditional probabilities. 

```{r probs myth}
# Between 27 and 29 minutes (the myth)
paste("The probability of waiting between 27 and 29 minutes for the ferry is",round(sum(analysis_data$wait_times_1min >= 27 & analysis_data$wait_times_1min <= 29.9)/dim(analysis_data)[1],2))
paste("The probability of waiting between 27 and 29 minutes for the ferry is",round(sum(analysis_data$wait_times_2min >= 27 & analysis_data$wait_times_2min <= 29.9)/dim(analysis_data)[1],2))
paste("The probability of waiting between 27 and 29 minutes for the ferry is",round(sum(analysis_data$wait_times_3min >= 27 & analysis_data$wait_times_3min <= 29.9)/dim(analysis_data)[1],2))
```

Whether or not it's surprising, missing the ferry by a couple minutes has a small probability. Let's see if there are differences if we condition on hour of day.

```{r probs myth rush}
myth_hour <- analysis_data %>% 
  group_by(as.factor(hour)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 27 & wait_times_1min <=29.9)/n(),
            prob_2min = sum(wait_times_2min >= 27 & wait_times_2min <=29.9)/n(),
            prob_3min = sum(wait_times_3min >= 27 & wait_times_3min <=29.9)/n())
colnames(myth_hour)[1] <- "Hour"
print.data.frame(myth_hour)
```

Some interesting patterns emerge in the conditional probabilities. If you take 3 minutes to exit, you have a 12% chance of just missing the ferry by minutes.  Yet if you take 1 or 2 minutes to exit the terminal, you have only an 8% and 6% chance of missing the ferry by minutes at 3 am, respectively. If you're aiming to catch a ferry in the 4 am hour, you have a 10% chance of missing it if you take 1 or 3 minutes to exit the terminal. 

Let's check the probabilities conditional on weekday. 

```{r prob myth day}
prob_weekday <- analysis_data %>% 
  group_by(as.factor(weekday)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 27 & wait_times_1min <=29.9)/n(),
            prob_2min = sum(wait_times_2min >= 27 & wait_times_2min <=29.9)/n(),
            prob_3min = sum(wait_times_3min >= 27 & wait_times_3min <=29.9)/n())
colnames(prob_weekday)[1] <- "Weekday"
print.data.frame(prob_weekday)
```

Once again, the probabilities are quite low, with a range from 0.03 to 0.05. 

Next I'll check the probability of waiting 27-29 minutes (inclusive) by month.

```{r probs myth month}
probs_month <- analysis_data %>% 
  group_by(as.factor(month)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 27 & wait_times_1min <=29.9)/n(),
            prob_2min = sum(wait_times_2min >= 27 & wait_times_2min <=29.9)/n(),
            prob_3min = sum(wait_times_3min >= 27 & wait_times_3min <=29.9)/n())
colnames(probs_month)[1] <- "Month"
print.data.frame(probs_month)
```

The 3 minute exit time had the highest probabilities across the wait time and months -- 0.059 in January and 0.054 in February. Overall the probabilities conditional on month look worse for 3 minute exit time. 

## The nightmare wait: 30+ minute waits

Although rare, 30+ minute waits do happen. I'll break it down marginally and conditionally to understand when such waits are more likely to happen. 

```{r nightmare marginal}
paste("The probability of waiting 30+ minutes for the ferry is",round(sum(analysis_data$wait_times_1min >= 30)/dim(analysis_data)[1],2))
paste("The probability of waiting 30+ minutes for the ferry is",round(sum(analysis_data$wait_times_2min >= 30)/dim(analysis_data)[1],2))
paste("The probability of waiting 30+ minutes for the ferry is",round(sum(analysis_data$wait_times_3min >= 30)/dim(analysis_data)[1],2))
```

The probabilities are quite similar to those from the myth. So the area under the curve between 30 minutes and 59.9 minutes is roughly the same as the area under the curve between 27 and 29 minutes. 

I'll break it down by hour, weekday, and month now.

```{r}
nightmare_hour <- analysis_data %>% 
  group_by(as.factor(hour)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 30)/n(),
            prob_2min = sum(wait_times_2min >= 30)/n(),
            prob_3min = sum(wait_times_3min >= 30)/n())
colnames(nightmare_hour)[1] <- "Hour"
print.data.frame(nightmare_hour)
```
Something strange happens around 23:00! The probability of waiting 30+ minutes is roughly 28%.

```{r nightmare weekday}
nightmare_weekday <- analysis_data %>% 
  group_by(as.factor(weekday)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 27 & wait_times_1min <=29.9)/n(),
            prob_2min = sum(wait_times_2min >= 27 & wait_times_2min <=29.9)/n(),
            prob_3min = sum(wait_times_3min >= 27 & wait_times_3min <=29.9)/n())
colnames(nightmare_weekday)[1] <- "Weekday"
print.data.frame(nightmare_weekday)
```

No single day is routinely bad. The highest probability is Monday for 1 minute exit time. 

Last step is to check probabilities conditionl on month.

```{r nightmare month}
nightmare_month <- analysis_data %>% 
  group_by(as.factor(month)) %>% 
  summarise(prob_1min = sum(wait_times_1min >= 30)/n(),
            prob_2min = sum(wait_times_2min >= 30)/n(),
            prob_3min = sum(wait_times_3min >= 30)/n())
colnames(nightmare_month)[1] <- "Month"
print.data.frame(nightmare_month)
```

The month of April wasn't great for 3 minute exit time with a 7% chance of a nightmare wait. February boasts a 7% chance of a nightmare wait for 2 minute exit time. Based on the probabilities conditional on month, running might not be a bad thing.
